meeting_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "meeting-env"
version = "1.0.0"
description = "OpenEnv-compliant Meeting Decision Intelligence Environment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "MetaxScaler"},
]
dependencies = [
    "pydantic>=2.7.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]

[project.scripts]
server = "meeting_env.server.app:app"

[tool.setuptools]
include-package-data = true
packages = ["meeting_env", "meeting_env.server"]
package-dir = { "meeting_env" = ".", "meeting_env.server" = "server" }
